refactor: split into a Cargo workspace of three crates
Carve the flat crate into domain-screaming libraries with a thin binary, per the
dependency rule (frameworks at the edges, no cross-dependency between the two libs):

- plaza-api   (lib) — REST client, Socket.IO feed, auth/token storage, models.
- plaza-audio (lib) — Player, the PcmSource decode pipeline, StreamQuality, and
                       the MP3/Opus/HLS sources.
- plaza-tui   (bin) — config, app state + run loop, and the terminal UI; a thin
                       main.rs wires the two libraries together.

The monolithic error enum is replaced by a per-library thiserror Error (plus a
Result alias); the binary uses anyhow. The audio engine no longer depends on an
async runtime: failures are reported through an ErrorReporter callback the binary
adapts to its event channel, so plaza-audio is framework-free.

Shared dependency versions are pinned in [workspace.dependencies]. The previous
main.rs re-declared every module, which both violated the lib/bin split and made
library APIs look like dead code; consuming the libraries from the binary fixes
both.

All four gates pass: cargo fmt --check, clippy -D warnings, test, and doc.
