+++ Cargo.toml
@@ -1,58 +1,68 @@
-[package]
-name = "plaza-tui"
+[workspace]
+resolver = "2"
+members = ["crates/plaza-api", "crates/plaza-audio", "crates/plaza-tui"]
+
+[workspace.package]
 version = "0.1.0"
 edition = "2021"
+rust-version = "1.85"
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/jtieri/plaza-tui"
 
-[dependencies]
-# TUI
-ratatui          = "0.30"
-crossterm        = { version = "0.29", features = ["event-stream"] }
-ratatui-image    = "10"
+[workspace.dependencies]
+# Internal crates
+plaza-api = { path = "crates/plaza-api" }
+plaza-audio = { path = "crates/plaza-audio" }
 
-# Async
-tokio            = { version = "1", features = ["full"] }
-tokio-stream     = "0.1"
+# Async runtime
+tokio = { version = "1", features = ["full"] }
+tokio-stream = "0.1"
+futures-util = "0.3"
 
 # HTTP
-reqwest          = { version = "0.12", default-features = false, features = ["json", "stream", "blocking", "rustls-tls"] }
+reqwest = { version = "0.12", default-features = false, features = [
+    "json",
+    "stream",
+    "blocking",
+    "rustls-tls",
+] }
 
-# JSON / config
-serde            = { version = "1", features = ["derive"] }
-serde_json       = "1"
-toml             = "0.8"
+# Serialization / config
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+toml = "0.8"
 
-# Socket.io
-rust_socketio    = { version = "0.6", features = ["async"] }
+# Real-time
+rust_socketio = { version = "0.6", features = ["async"] }
 
-# HLS
-m3u8-rs          = "6"
+# Audio
+symphonia = { version = "0.5", features = ["aac", "mp3", "ogg", "isomp4"] }
+rodio = { version = "0.19", default-features = false, features = ["symphonia-all"] }
+opus = "0.3"
+m3u8-rs = "6"
 
-# Audio decoding
-symphonia        = { version = "0.5", features = ["aac", "mp3", "ogg", "isomp4"] }
-# Opus decoder (symphonia has none). FFI to libopus, statically vendored/built
-# from source so released binaries need no system library.
-opus             = "0.3"
+# TUI
+ratatui = "0.30"
+crossterm = { version = "0.29", features = ["event-stream"] }
+ratatui-image = "10"
+image = "0.25"
 
-# Audio output
-rodio            = { version = "0.19", default-features = false, features = ["symphonia-all"] }
+# Storage / platform
+keyring = { version = "3", features = ["sync-secret-service", "crypto-rust"] }
+dirs = "5"
 
-# Secure storage
-keyring          = { version = "3", features = ["sync-secret-service", "crypto-rust"] }
+# Diagnostics
+tracing = "0.1"
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 
 # Error handling
-anyhow           = "1"
-thiserror        = "1"
+thiserror = "1"
+anyhow = "1"
 
-# Utilities
-clap             = { version = "4", features = ["derive"] }
-dirs             = "5"
-chrono           = { version = "0.4", features = ["serde"] }
-tracing          = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-bytes            = "1"
-image            = "0.25"
-futures-util     = "0.3"
+# CLI / time
+clap = { version = "4", features = ["derive"] }
+chrono = { version = "0.4", features = ["serde"] }
 
-[dev-dependencies]
-wiremock         = "0.6"
-tempfile         = "3"
+# Dev / testing
+wiremock = "0.6"
+tempfile = "3"
