+++ crates/plaza-api/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "plaza-api"
+description = "Nightwave Plaza REST, real-time, and auth client"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+reqwest.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+rust_socketio.workspace = true
+keyring.workspace = true
+dirs.workspace = true
+chrono.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+thiserror.workspace = true
+
+[dev-dependencies]
+tokio = { workspace = true }
+wiremock.workspace = true
