+++ crates/plaza-audio/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "plaza-audio"
+description = "Streaming audio engine for Plaza TUI: MP3, Opus, and HLS/AAC playback"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+symphonia.workspace = true
+rodio.workspace = true
+opus.workspace = true
+m3u8-rs.workspace = true
+reqwest.workspace = true
+serde.workspace = true
+tracing.workspace = true
+thiserror.workspace = true
