+++ crates/plaza-tui/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "plaza-tui"
+description = "A terminal client for Nightwave Plaza, the vaporwave radio station"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+default-run = "plaza-tui"
+
+[[bin]]
+name = "plaza-tui"
+path = "src/main.rs"
+
+[lib]
+name = "plaza_tui"
+path = "src/lib.rs"
+
+[dependencies]
+plaza-api.workspace = true
+plaza-audio.workspace = true
+
+tokio.workspace = true
+tokio-stream.workspace = true
+futures-util.workspace = true
+
+ratatui.workspace = true
+crossterm.workspace = true
+ratatui-image.workspace = true
+image.workspace = true
+
+reqwest.workspace = true
+serde.workspace = true
+toml.workspace = true
+dirs.workspace = true
+chrono.workspace = true
+
+clap.workspace = true
+tracing.workspace = true
+tracing-subscriber.workspace = true
+anyhow.workspace = true
+
+[dev-dependencies]
+tempfile.workspace = true
