+++ crates/plaza-tui/src/tui/views/favorites.rs
