+++ crates/plaza-tui/src/tui/views/now_playing.rs
