--- src/audio/mod.rs
@@ -1,7 +1,0 @@
-pub mod hls;
-pub mod pcm;
-pub mod player;
-pub mod sources;
-pub mod ts;
-
-pub use player::Player;
